# product-api/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  GenericError:
    description: GenericError is a generic error message returned by a server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/firkhraag/product-api/internal/handlers
  Product:
    description: Product defines the structure for an API product
    properties:
      decription:
        description: the description for the product
        maxLength: 10000
        type: string
        x-go-name: Description
      id:
        description: the id for the user
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        description: the name for the product
        maxLength: 255
        type: string
        x-go-name: Name
      price:
        description: the price for the product
        format: float
        minimum: 0.01
        type: number
        x-go-name: Price
      sku:
        description: the SKU for the product
        pattern: '[a-z]+-[a-z]+-[a-z]+'
        type: string
        x-go-name: SKU
    required:
    - id
    - name
    - price
    - sku
    type: object
    x-go-package: github.com/firkhraag/product-api/internal/model
  ValidationError:
    description: ValidationError is a collection of validation error messages
    properties:
      messages:
        items:
          type: string
        type: array
        x-go-name: Messages
    type: object
    x-go-package: github.com/firkhraag/product-api/internal/handlers
info:
  description: Documentation for Product API
  title: Product API
  version: 1.0.0
paths:
  /{id}:
    get:
      description: Returns a product
      operationId: getProduct
      parameters:
      - description: The id of the product
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/productResponse'
        "404":
          $ref: '#/responses/errorResponse'
        "500":
          $ref: '#/responses/errorResponse'
      tags:
      - products
  /products:
    get:
      description: Returns a list of products
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
        "500":
          $ref: '#/responses/errorResponse'
      tags:
      - products
    post:
      description: Adds a product
      operationId: addProduct
      responses:
        "200":
          $ref: '#/responses/noContentResponse'
        "400":
          $ref: '#/responses/errorValidation'
        "422":
          $ref: '#/responses/errorResponse'
      tags:
      - products
    put:
      description: Updates a product
      operationId: updateProduct
      responses:
        "200":
          $ref: '#/responses/noContentResponse'
        "400":
          $ref: '#/responses/errorValidation'
        "404":
          $ref: '#/responses/errorResponse'
        "422":
          $ref: '#/responses/errorValidation'
      tags:
      - products
  /products/{id}:
    delete:
      description: Deletes a product
      operationId: deleteProduct
      parameters:
      - description: The id of the product
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/noContentResponse'
        "404":
          $ref: '#/responses/errorResponse'
      tags:
      - products
produces:
- application/json
responses:
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
  errorValidation:
    description: Validation errors defined as an array of strings
    schema:
      $ref: '#/definitions/ValidationError'
  noContentResponse:
    description: This API endpoint returns no content
  productResponse:
    description: This API endpoint returns a product
    schema:
      $ref: '#/definitions/Product'
  productsResponse:
    description: This API endpoint returns a list of products
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
schemes:
- http
swagger: "2.0"